feat: support Windows hosts
Connects over the podman machine named pipe (queried via podman machine
inspect, falling back to the default machine's pipe) and moves inline
secret/config staging into engine::staging with per-platform behavior:
unix keeps the verified 0700 XDG staging exactly as before; Windows
stages under the per-user temp directory, whose default ACLs restrict
access to the owning user, and warns when compose mode/uid/gid fields
cannot apply. libc becomes a unix-only dependency.
