+++ Cargo.toml
@@ -32,12 +32,14 @@ bollard = "0.21"
 clap = { version = "4", features = ["derive", "env"] }
 indexmap = { version = "2", features = ["serde"] }
 futures = "0.3"
-libc = "0.2"
 tar = "0.4"
 flate2 = "1.0"
 bytes = "1"
 walkdir = "2"
 notify = "8"
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
 
 [dev-dependencies]
 tempfile = "3"
